% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{SmoothRasterTS}
\alias{SmoothRasterTS}
\title{Smooth Raster Time Series}
\usage{
SmoothRasterTS(rast, ...)
}
\arguments{
\item{rast}{SpatRaster. A multi-layer raster object where each layer
represents a time point in the series.}

\item{...}{Additional arguments passed to the \code{SmoothTS} function, such as
\code{method}, \code{n}, \code{weights}, \code{alpha}, and \code{n_init}.}
}
\value{
SpatRaster. A multi-layer raster object with smoothed time series for
each pixel.
}
\description{
Applies a smoothing operation to each pixel's time series in a multi-layer
raster using one of three methods: mean (moving average), weighted moving
average, or exponential moving average.
}
\details{
Assumes that the input raster has multiple layers, where each layer
represents a time point in the series. The function applies the specified
smoothing method to the time series of each pixel across the layers and
returns a new raster with the smoothed time series for each pixel.
}
