% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{SmoothRasterTS}
\alias{SmoothRasterTS}
\title{Smooth Raster Time Series}
\usage{
SmoothRasterTS(
  rast,
  ...,
  cores = 1L,
  filename = "",
  overwrite = FALSE,
  wopt = list()
)
}
\arguments{
\item{rast}{SpatRaster. A multi-layer raster object where each layer
represents a time point in the series.}

\item{...}{Additional arguments passed to the \code{SmoothTS} function, such as
\code{method}, \code{n}, \code{weights}, \code{alpha}, and \code{n_init}.}

\item{cores}{Integer. Number of CPU cores to use for parallel processing.
Default is 1. Don't change this unless you know what you're doing.}

\item{filename}{Character. Optional filename for writing the output raster.
If provided, the output raster will be written to this file.}

\item{overwrite}{Logical. If TRUE and \code{filename} is provided, allows
overwriting the existing file; optional.}

\item{wopt}{List. Optional list of additional arguments to pass to
\code{terra::writeRaster}.}
}
\value{
SpatRaster. A multi-layer raster object with smoothed time series for
each pixel.
}
\description{
Applies a smoothing operation to each pixel's time series in a multi-layer
raster using one of three methods: mean (moving average), weighted moving
average, or exponential moving average.
}
\details{
Assumes that the input raster has multiple layers, where each layer
represents a time point in the series. The function applies the specified
smoothing method to the time series of each pixel across the layers and
returns a new raster with the smoothed time series for each pixel.
}
\examples{

rast <- terra::rast(nrow = 1, ncol = 2, nlyrs = 5)
terra::values(rast) <- matrix(rep(1:5, each = 2), nrow = 2)
out <- SmoothRasterTS(rast, method = "mean", n = 3)

}
